Changed various data types to int; because Java collections/arrays do not support more than Integer.MAX_VALUE elements

+++ src/constants/Constants.java
@@ -1,8 +1,10 @@
 package constants;
 
 public class Constants {
-    public static long N = 10000; //Total size of the file (probably should be read)
-    public static long M = (long) Math.sqrt(N);
-    public static long g = 5; //Number of Column Groups
-    public static long groupLength = M/g; // Length of a ColumnGroup
+    
+    //The production file has 400,000,000 lines, which will fit in an int
+    public static int N = 10000; //Total size of the file (probably should be read)
+    public static int M = (int) Math.sqrt(N);
+    public static int g = 5; //Number of Column Groups
+    public static int groupLength = M/g; // Length of a ColumnGroup
 }
